Better handle loading palette files with >4096 columns
Attempting to set the number of columns will immediately
start allocating QMap instances for every column.

For maliciously crafted files, this can take a long time
and waste a whole lot of memory, forcing the user to kill
krita or waiting for the operating system to OOM-kill it.

On invalid column count, the parser will behave as if no
column count had been present at all.