Add color space import/export compatibility tests
Before there was no way of checking if an image saved in one color space
is opened in the same color space. For example tiff import can assume
things about color space of the image that isn't necessarily true.
This commit adds a possible test to use to check it.