Fix availability of vectors on cloned vector layers (again)
KoSelection emits signals using signal compressor, which
is placed in the GUI thread (because of timers), therefore
it might receive queued signals. And for parsing queued
signals it needs an event loop...

So we just ensure that KoShapeManager and KoSelection belong
to the GUI thread, where an event loop is surely present.

BUG:408028