Fix 0 clones in animation rendering
Some users encountered message:
"Memory limit has been reached. The number of frames saved simultaneously
is limited to 0"
in some special circumstances related to lack of memory.
In this case Krita would infinitely wait for the user to click Cancel
pretending to be rendering frames (but not doing so).

This commit fixes this behaviour by ensuring the negative calculated
memory amount is not propagated further.