Fix 0 fuzziness in Similar Color Selection Tool
In Contiguous Color Selection Tool the fuzziness slider goes from 1 up.
In Similar Color Selection Tool in the code there is an explicit check
for fuzziness = 1 with an assumption there is no fuzziness lower than 1.
However the slider didn't conform to that before this commit.
This commit fixes this situation, now the fuzziness slider for Similar
Color Selection Tool goes from 1 up.