Fix merging layers with not active selection masks
Before this commit, there was an assumption in the merging code
that a layer can have only one selection mask (which can be
gathered using ->selectionMask() function). This caused all non-active
selection mask to disappear during/after the merging.
This commit fixes that behaviour by taking all children of the merged
layer that happen to be KisSelectionMasks into account.

Note: it doesn't mean all of the selection masks will be preserved;
for now only the immediate child nodes of the selected layers are
considered.

BUG:402070