Fix cloning not-opacity scalar keyframe channels
Before this commit, duplicating a layer with keyframe channels
that weren't specifically an opacity keyframe channel resulted in
loss of those channels. This commit provides an universal solution for
all scalar kyeframe channels available.

At this point there is no way of making a scalar keyframe channel
except for opacity keyframes unless you make a change in the code and
compile it yourself (see animation_docker.cpp, line 508), so there
is/should be no change in behaviour for users. However if you enable
animated transform masks, this become crucial for both duplicating
layers and saving the file with keyframes (it doesn't work for now,
but without this commit it most probably wouldn't).