Implement "Software Renderer" option for OpenGL engine
On old GPUs that support D3D<=10.0 we cannot use Angle for emulating
openGL, therefore we should use some software implementation (or try
using D3D9 with its openGLES 2.0). This patch adds an option to select
software rasterizer WARP, created by Microsoft, which is a part of
DirectX 11 and should be avaailable on all system with Windows 7 and
newer.

CCBUG:408872