Fix switching profile to default in New file dialog
Before this commit if you first selected the color model,
then color profile, then color depth, the color profile would switch to
default for this color depth no matter that the user just set it
manually. This is mostly the problem if someone is for example using
Rec2020 profile for HDR images and switches between 32 and 16 bit
color depths.

Behaviour now:
- by default color profile changes to the default one that suits
the color model and color depth
- if the user sets the profile manually, it is remembered
and if it changes color depth, the profile doesn't change
(keep in mind that "(Default)" suffix is still attached
to the default profile, so it's easy to find in case the user wants it
- if the user changes the color model, the default behaviour
(switching profile automatically) is restored until the user
sets up the profile maually again.

BUG:406700