Fix crash on CMYK profiles not updated
Before this commit there was a crash if the default (last used)
color model was not CMYK, but the user wanted to create a CMYK
file. Color profiles didn't get updated, so Krita tried to create
a document with an incompatible color profile which resulted in crash.
This commit ensures the correct color model is used.

BUG:409916