Fix update brush size using shortcuts or drag
Before this commit, brush size slider wasn't updated.
This patch fixes this behaviour.

Looks like at the same time m_resources *doesn't* contain
a specific key while m_updateMediators does and needs
this connectResource thing.
The key in question is 6007 with KisPaintOpPresetSP value,
which means it's "CurrentPaintOpPreset".

This probably needs some more investigation. This fix will hide
the issue for users though. It is based on revert of the commit
7c2fce1dbdf005bfb33e687f6243f061b606ca5b.

BUG:408331