Use regular asserts in KisDocument::slotChildCompletedSavingInBackground
Use KIS_ASSERT instead of KIS_SAFE_ASSERT in
KisDocument::slotChildCompletedSavingInBackground. When any of these
conditions triggers, it means that something is really wrong and the
saved file may be corrupt. On systems with safe asserts hidden (e.g.
Windows), there is no feedback to the user that something bad happened
(apart from the fact that the document is still modified after save).