Add subcategories to ‘appdata’ and .desktop files
This adds the subcategories ‘2DGraphics’ and ‘RasterGraphics’ to the
metadata files (the main category ‘Graphics’ was already present).
The official list of possible subcategories can be found at:
https://standards.freedesktop.org/menu-spec/latest/apas02.html

With this change, Krita should appear in the ‘Painting and Editing’
(sub)category in Plasma Discover (along with similar applications,
like GIMP and MyPaint).