Disambiguate various ‘New’/‘New...’ strings for translators
For some languages, the translation of the word ‘New’ depends
on which type of ‘new’ object that is created (e.g. on the gender
of the word ‘xxx’ in the string ‘New xxx’). This commit adds
message context (msgctxt) to disambiguate the various ‘New’/‘New...’
strings.

It also fixes a few inconsistencies in the use (or not) of ‘...’
and the placement of the ‘&’ accelerator marker.

(cherry picked from commit fb58616ff2a182dab22ee3bb48ef2c8d374e4bb2)
(cherry picked from commit 871875e529852fc7d2509203eb6b0f408aa2c8ae)