Use translation of unnamed (‘Untitled’) colours
All of the colour palettes included in Krita that have unnamed colours,
have actually the colour names set to ‘Untitled’ in the .gpl files,
instead of having empty names (I’m not sure it’s valid to have empty
names, though). So the code that tried to show unnamed colours as the
translation of the string ‘Untitled’, did not succeed. Now, colours with
empty names or the name ‘Untitled’ are treated the same, so that the
i18n() call always succeeds for unnamed colours.