Show publication time of RSS news in user’s preferred short datetime format
The publication time of the news items on Krita’s welcome page is now
shown in the user’s timezone and in their preferred short datetime format.
The timezone of the datetime in the RSS feed is now also taken into account
when parsing the publication date (the old code basically assumed that the
user lived in the timezone the RSS feed generator was set up to use, which
at the time of writing was UTC).

Note that old code only displayed the date; the new code also displays
the time (typically hh:mm).