Fix updates in the new transform tool
Batch markers must not be compressed. I thought that returning a fake
LoD would ensure it, but it is not the case. If we have multiple "start"
(or "end") markers, they should not be compressed as well. For this
purpose, I introduced KisUpdateInfo::canBeCompressed(), which returns
false for all the marker info objects.