Show/store translated name of the ‘Current’ calligraphy profile
The name of the ‘Current’ calligraphy profile should now be shown
as translated. It’s also *stored* with the translated name
(i.e. the translation of the string ‘Current’). This doesn’t seem
optimal, but the rest of the existing code seems to depend on this.
And the two other profile names (‘Mouse’ and ‘Graphics Pen’) are
also stored with the translated names (meaning that they don’t
change when/if the translation is changed).