Add message context for the four transformation types
The message context is useful for the translators (for disambiguation).

Also removed the accelerator keys from two of the strings. Only two
out of the four strings had accelerators, the accelerators were in
conflict with the accelerators used in the menu bar, and it was not
possible to add non-conflicting accelerators for all four types of
transformation.