Remove misleading ‘px’ suffix for ‘move amount’ shortcut setting
The ‘Amount’ input field for the move tool had a ‘px’ suffix.
But the unit used is chosen from a different dropdown field,
so the input field could end up saying ‘0.3527px’ ‘Millimeters (mm)’,
which didn’t make sense. Also updated the tooltip to reflect that
the unit can be chosen by the user.