Remove the obsolete xmlgui files from the plugins
Our plugins weren't xmlguiclients anymore to begin with, so these
files couldn't do anything whatsoever. I've tried to restore this
functionality to KisActionPlugin, that wasn't a big success since
merging menus together gives unpredictable results by default. And
if we'd add all kinds of merge points in the main menu, we might
as well not bother.