Fix transform worker tests failing because of NN
Tests for transform worker were failing because the result files
contained results from the old version of NN/Box filter.
Since I changed it a bit in 57c3194c1c8aaf9b17fd8f080533610c38a626a7
the old result files were no longer valid.
This commit updates the result files.

Also since the execution of the filter on one of the images
resulted in a transparent pixel in the location that the previous filter
had a non-transparent pixel, it changed the size of the resulting image
so it didn't pass the test partially because of the change in size.
Since other tests don't check the size, I removed the checks in this
test, too. (Because of the transparency in the test image, it is possible
that the resulting image will have a different size even though it is
not scaled in this dimension.)

BACKPORT:krita/4.2