Fix Strength not used in Rotation - Fuzzy Dab
Before this commit, if you chose Fuzzy Dab sensor
in Rotation option, the value of the Strength slider wouldn't
affect the resulting rotation of the brush tip.
This commit fixes that behaviour: now if you choose very low
Strength and the first curve from the template ones,
the brush tip will rotate only slightly (before it would rotate
within the full 360* range).

* Fuzzy Dab is 'additive' kind of sensor, so it's value would be put
in 'realAdditivePart' in rotationLikeValue(). 'constant' contains the
value of 'Strength' slider. To get it working together,
'realAdditivePart' had to be multiplied by the 'constant' value.

* HSV option just ignored the problem and multiplied it by the value of
the slider later. It couldn't be done for Rotation because Rotation
takes the canvas rotation into account (see rotationLikeValue).
Therefore it had to be done inside rotationLikeValue function, not
later. Therefore the multiplication in HSV option had to be disabled
to not double the effect.

BUG:376179
BACKPORT:krita/4.2