Add support for fractions with 2 decimals numbers to the FloatColorInput
Fractions which resulted in doubles with 2 decimal numbers or less,
could not be converted. The fraction would only be converted if the
rounded value differs from the full fraction, which is not the case if
they are the same.

This check is not needed anymore, since the slider won't trigger the
input field anymore. Which would result in the early convertion.