Optimize LoG generation Magneti Selection tool
For Colorize Mask we set kernel size to be (4 * radius + 1) to ensure
that the watershed algorithm stops **not reaching** the border of the
color blob. In the selection tool, it is the opposite. We should stop
right at the edge. Therefore, the kernel size should be (2 * radius + 1)
and should not include the wrapped art of the LoG surface.