Fix lost scroll events when tablet is in the proximity
If there is a really heavy flow of events, the OS will start compressing
wheel events into bigger deltas. So we should do multiple zoom actions,
not one. It might fix the problem on Windows, but it will *not* fix the
problem on Linux, because on Linux XCB just drops all the MOTION events
for the mouse-wheel subdevice and sends DEVICE_CHANGED events instead.
And there is no easy way to generate QWheelEvent from DEVICE_CHANGED
event.

CCBUG:410940