Add size hints to tablet tester widget (for better dialogue size)
The tablet tester dialogue’s ‘preferred’ size is being overridden
by KoDialog, resulting in a very small dialogue box, with a tiny
painting area for testing the tablet. Now we add a sizeHint for this
area, so that the dialogue automatically gets a reasonable size.

The width and height values for the sizeHint was chosen so that
they correspond to the size the widget would have had if KoDialog
didn’t override the preferred window size.