Change relative widths of the tablet tester and status messages widgets
The status message textedit field was too narrow to show a status
message without breaking it over several lines. The relative widths of
the tablet tester painting area and the textedit field have therefore
been changed to make the latter wider, so that there’s enough horizontal
space for the status messages. The sizeHint for the tablet tester area
also had to be changed to make the dialogue wide enough.