Implement transformation of the Grid when image is transformed
Basically, we implement a fake layer (KisDecorationsWrapperLayer),
which accepts all the transformations and tries to transform decorations
(only grids are implemented atm).

There is a tricky thing: tranformations are undoable, but modifications
to grids are not! Therefore we should be really careful about transforming
grids. We cannot just save their state, like we plan to do for all the
vector objects. Instead, we should apply inverted transformation during
undo operation.

CCBUG:331789