Replace cache with a new variable in curve option
Before this commit, "Use the same curve" (precisely,
"Share curve across all settings") was done using
(1) rewriting values in sensors to the selected curve
(2) cached curves for all the sensors in case the user changes
their mind and wants the previous ones.

This behaviour brings confusion and is hard to debug, hence
the change.

This commit adds a new separate variable that keeps the
"common curve", so all sensors has their own curves.
There is no need for caching or restoring, it just needs
to be clear that when the Use The Same Curve checkbox is
checked, the value from the common curve needs to be taken
instead of the curve from the sensor in question.