Fix shearing Ellipse and Concentric Ellipse assistants
Ellipses are rather tricky to transform, they main axes change
non-uniformly during shear operation, so a special handling is needed.

This approach uses matric representation of an ellipse equation to
get new axes values. We just decompose the matrix into diagonal
matrix of eigen values and a orthogonal rotation matrix using Eigen
library. The solution might be not the easiest one, but it works :)