Make HTML markup in search field tooltips work
The tooltips for some search text fields (used for example
in the brush preset widget) used HTML code, but didn’t
start with an HTML tag, so it was interpreted as plain
text.

Since the strings had to be changed anyway, this commmit
also has some other minor improvements to the strings,
e.g. it removes some unnecessary manual line breaks
(which caused bad formatting).

The two very similar (and long) strings are also
‘deduplicated’, by extracting out their common text to a
separate string, which makes them easier to translate.