Possibly fix File Layer loading and comics manager page updating on windows.
Windows has a pretty slow file system, which means when we check whether
our updated file exists, windows hasn't gotten to renaming it yet, making
Krita think there's no file at all...

So this puts a little bit of a delay in our two checks of whether the file still
exists and needs to be readded to the filesystemwatcher again.

This currently works as expected on Linux, now we need to test if it works
on Windows.

CCBUG:410409
CCBUG:389544


(cherry picked from commit ff338de7354fc170645af1a435267d1024546654)