Make use of LCMS' copy alpha flag instead of copying alpha ourselves.
This should make color management related transforms a bit faster, as
well as simplifying the code. Given an enum in the depths of the pigment
jungle is being changed, it takes long to compile.

boud thought there was more alpha handling that needed to be changed,
but this is the only place I can see with a workaround...

Ref T3503