Fix a crash when loading JPEG image with a weird color profile
The profile in question has a weird RGB->XYZ transformation matrix,
which is not invertible. Therefore, this profile cannot be used as
a workspace color profile and we should convert the image to sRGB
right on loading.

LCMS doesn't have a separate method for checking if conversion matrix
is invertible, therefore we just try to create a simple transformation,
where the profile is both, input and output. If the transformation
is created successfully, then this profile is probably suitable for
usage as a working color space.

BUG:410864