Implement some tests.
These test whether the pixels are loaded correctly, the blending modes,
the layer configuration and the clipping groups are configured right.

Then there's two of the standard opening tests.

Clipping groups test fails because I made a mistake in the code,
one of the opening tests fails, I suspect because Krita's not yet
recognizing the mimetype right...

Ref T11330