Fix presets with random offset of texture marked dirty
Before this commit, if a preset had at least one random offset (X or Y)
in texture settings (and texture enabled), it was marked dirty as soon
as the user used this brush on the canvas. That was happenning because
random offset just means that the offset of the texture is set
to a new random one after every stroke. Setting a property to a new value
marked preset as dirty.
This commit ensures that in this case te preset is not set to dirty
(it just disables dirty notifications while the new offset is set).

BUG:406427