Test patch that "fixes" HSL transfgormation for white and black images
Basically, the transformation is substituted with HCL transformation,
but with clever sliding on the edges of the rhombus. I need people's
input, whether the transformation is still useful for painters.

Now it looks almost like "Hue/Saturation/Luma" (it also uses chroma
instead of saturation), but with a different approach to calculation
of 'L' channel :)

CCBUG:411863