Fix updating current shape color when undo/redo
shapeChanged() is also called when the selection emits
shapeContentChanged(), that is, when the list of shapes is still
the same, but some properties of shapes are changed. In such
a case we should still update the controls.

The offending code was originally intended to fix bug 404975,
but it seems like the bug doesn't happen anymore.

CCBUG:404975