Prevent duplicate svg symbols in the vector library docker.
This was caused by the fact that right now, whenever svg use elements are used,
instead of loading it into something or the other, it just reparses the
definition the use is referring to. This is a robust fallback, but also led to
duplicate entries in the vector library docker.

Now the parser uses a somewhat similar system as the one in use for other def
citizens like the gradients, clip paths and masks, filters and markers, and will
not allow duplicate ID entries anymore.

This is also consistent with the SVG spec.

BUG:406634