CID 35534: Uninitialized pointer field (UNINIT_CTOR)
18. uninit_member: Non-static class member _lastOpacityData is not
initialized in this constructor nor in any functions that it calls.

2. uninit_member: Non-static class member dstRowStart is not initialized
in this constructor nor in any functions that it calls.

4. uninit_member: Non-static class member dstRowStride is not initialized
in this constructor nor in any functions that it calls.

6. uninit_member: Non-static class member srcRowStart is not initialized
in this constructor nor in any functions that it calls.

8. uninit_member: Non-static class member srcRowStride is not initialized
in this constructor nor in any functions that it calls.

10. uninit_member: Non-static class member maskRowStart is not initialized
in this constructor nor in any functions that it calls.

12. uninit_member: Non-static class member maskRowStride is not
initialized in this constructor nor in any functions that it calls.

14. uninit_member: Non-static class member rows is not initialized in
this constructor nor in any functions that it calls.

16. uninit_member: Non-static class member cols is not initialized in
this constructor nor in any functions that it calls.

(cherry picked from commit 2354e8f3679ce51f55197e19abe28f4f9c2185b3)