Work around a crash in Qt when converting a QDomDocument to a string
In Qt 5.14, if that is done in a thread which is not the gui thread,
QRegularExpression won't have a QThreadStorage and will crash. It is
weird, though, that Qt's exit handlers clear static variables in
non-gui threads.

I'm not sure what changed in Qt: the exit handler or QRegularExpression.

Since we save the tags on every change, the saving on exit isn't strictly
needed, so we can remove that to work around this issue in Qt.

BUG:410867