Fix Rec2020 display format
This commit contains changes to smpte shader to fix the black screen
in Rec2020 display format on Windows.
Before this commit it was broken because of two things:
- r=pow(x, y) in ANGLE is compiled into t=log(x); t=t*y; r=e^t
which in case of 0 resulted in undefined behaviour, in this case...
alpha = 1.0, even though the line was: pow(0, 1), qhich means
the correct result should be 0.0, not 1.0.
- changed order of painting UI and canvas; before it was first UI,
then canvas. After the order was reversed, so the 0.0 alpha started
to be crucial for canvas to be shown from underneath the UI.

Co-Authored-By:  Dmitry Kazakov <dimula73@gmail.com>

BUG:410918
(cherry picked from commit 2f976a938a8f8e7721bc5d9393289ff98b55b552)