Fix wrong aspect ratio on SVG Load
We incorrectly treated the absence of preserveAspectRatio
as set to "none". Svg specification states the initial value
of this property should be "xMidYMid meet". in case there
is no property we now use the initial value instead of
ignoring.
    
Updated TestSvgParser cases to use "none" so the resulting
shape corresponds to the old behaviour.
    
BUG:407425