Fix an assert when force-autosaving the image right during the stroke
When we make a clone of a shape layer, we must ensure that no updates
are initiated after cloning the image. We used to block updates at
the level of KisShapeLayerCanvas, but it works only when we save/clone
the image from the GUI thread. When we save it from the worker thread,
KoShapeManager queues the event into the GUI events queue, so an update
comes asynchronously.

To resolve this issue, the patch moves locking from KisShapeLayerCanvas
to KoShapeManager. We must ensure that there is no compressors nor queues
between the adding code and blocking code.

BUG:412835