Rewrite "Auto Precision" option
The patch basically removes old implementation of auto-brush and
substitutes it with simple heuristics:

1) If brush size is below 30px, use full precision (level 5)
2) If brush has a texture, scatter, mirror, rotation or airbrush
   option, then disable subpixel precision (level 3)

The heuristics are chosen this way to make sure that all the older presets
work fine. The problem is that auto-precision option did nothing for
several previous versions of Krita (it just returned precision 5 all
the time), therefore brush creators bascially couldn't configure it
properly. The new implementation makes sure that most of the presets
with auto-precision will still have precision level 5, like it used to
be before auto-precision "fix".

The patch does *not* add any UI strings (only removes a few). So it
is safe to be backported to 4.2.

BUG:413085

# Conflicts:
#	plugins/paintops/libpaintop/kis_brush_selection_widget.cpp
#	plugins/paintops/libpaintop/kis_brush_selection_widget.h
#	plugins/paintops/libpaintop/kis_dab_cache_base.cpp