Implement KisAbrStorage and its tests
This commit implements KisAbrStorage allowing the user
(or for now, tests) to load ABR files as storages for
resources.

Notes on the code:
- AbrResourceIterator uses iterator++ and --, because I couldn't
find Java-like QMap iterator (as they are for QList etc.), there
were only STL-like, which means hasNext() had to be implemented
in this hacky way
- I changed constness of next() function since it looks like lying
otherwise
- constness of hasNext() had to be faked... not sure why we want them
to be const if it's not true anyway? (the way to do it was taken from
KisBundleStorage)
- I added .abr taken from:
https://www.deviantart.com/mar-ka/art/brushes-186829534
license looks permissive enough
- I had to change QMap in abr_brush_collection to be
QSharedPointer, so using operator [] is very ugly now,
but that hopefully isn't a problem since it is only needed
a few times