Fix cloning of the image with overlay selection mask
When cloning the image with a copy-ctor it should have exactly
the same state as before cloning. It includes the state of the
overlay selection mask. It guarantees that there will be no async
jobs running right after the operation.

When saving, we don't want the overlay mask be saved into the file,
therefore we should run a separate action to disable it.

BUG:412747

# Conflicts:
#	libs/image/kis_image.cc