Remove KisPSDLayerStyleResource
Asl is supposed to be loaded through KisAslStorage,
and every layer style should be a separate resource.
This commit removes KisPSDLayerStyleResource altogether
since it was a collection of layer styles that was treated as one
resource instead of a storage. This commit also removes loading and
saving layer styles from and to .psd and .kra files, because the
author didn't know at the time of writing what is the correct
approach to implement it.

Things that needs implementing:
- loading layer styles in kis_kra_loader.cpp
- loading layer styles in psd_loader.cpp
- saving layer styles in kis_kra_saver.cpp
- loading layer styles in kis_dlg_layer_style.cpp from .asl files