Add virtual KoResource::resourceType()
Thanks to that addition, from now on resources are
self-aware which type are they.
Since it's virtual, it allows the programmer to check
the type in a better way than just asking "is it this type"
based on the base classes for all resource types.
Compare with KisWindowLayoutResource and
KisSessionResource.