Fix undo of replacing vector selection
When redoing KisTakeAllShapesCommand, we should keep the
selection projection, but when undoing it, we should start a
full-featured update to get a correct canvas state. It is needed
because pixel selection is used as a projection in shape-mode, so
undo/redo for it is not valid.

BUG:412808


(cherry picked from commit 7a709c5480aa24bb693a937feb8465f5e0547c4e)