Fix outline in Move Tool, when a layer has a Transform Mask
We need to take layer's masks' change rect into account. We cannot
use normal changeRect() call, because it never shrinks the update rect
and adds projection device bounds (obviously, because it is supposed
to render stuff and not just shwo a decoration).

That is why the patch introduces KisLayer::tightUserVisibleBounds(), which
calculates tight bounding rect of the layer, taking all the masks into
account.

BUG:410163