Fix comparison of double spin box
When the value of a QDoubleSpinBox is changed it will be round down to the amount of decimals that are assigned to that spin box. The KisDoubleParseSpinBox would check the new unrounded value with the current rounded value, which resulted in an incorrect comparison.

(cherry picked from commit 8bc2965db84224938f5bb3b791a04b98fe70e1a7)