Add KisTagModel to KisTagChooserWidget
For now, KisTagChooserWidget only takes all the tags from the model
and populate its' tags list using tag names from the model.
Later it will be needed to replace the custom tags and readOnlyTags
lists with the model, so that all changes to the model are automatically
reflected in the view.