Use KisTagSP in signals in tagging widgets
This is a commit with a lot of changes. Most of them
are incomplete. It's a result of experimentation
in the middle of a huge rewrite and shouldn't be considered
complete or whole in any type of circumstances.
It does compile and shouldn't crash, the only
area that was changed is tagging.

Changes include:
- changing all QString& signals to KisTagSP (unless not
applicable, for example in case of the filter widget)
- commenting out a lot of code
- adding a lot of debug